# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Propolis Fulfillment API",
        "description": "Dokumentasi resmi layanan backend Propolis. Semua response menggunakan format {success, data, message}.",
        "contact": {
            "name": "Tim Propolis",
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "API Root"
        },
        {
            "url": "http://localhost:8000/api"
        }
    ],
    "paths": {
...
        "/products/{productId}/batches": {
            "get": {
                "tags": [
                    "Batches"
                ],
                "summary": "Daftar batch untuk produk",
                "operationId": "3cdef7944210edf969612d50eda4b296",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar batch"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Batches"
                ],
                "summary": "Tambah batch baru / restock",
                "operationId": "58315c63c52c24504a80010bd07218e0",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "batch_number",
                                    "qty_initial"
                                ],
                                "properties": {
                                    "batch_number": {
                                        "type": "string"
                                    },
                                    "qty_initial": {
                                        "type": "integer"
                                    },
                                    "expiry_date": {
                                        "type": "string",
                                        "format": "date",
                                        "nullable": true
                                    },
                                    "purchase_price": {
                                        "type": "number",
                                        "format": "float",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Batch berhasil ditambahkan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/orders/{id}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Detail order milik pelanggan",
                "operationId": "f088d34a2737782215f435f854e31de5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail order beserta alokasi batch"
                    },
                    "403": {
                        "description": "Tidak berhak mengakses order ini"
                    },
                    "404": {
                        "description": "Order tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/orders/{id}/upload-proof": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Upload bukti pembayaran",
                "operationId": "411db1287f08c3903f826f0782f27596",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "bukti"
                                ],
                                "properties": {
                                    "bukti": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Bukti pembayaran tersimpan"
                    },
                    "404": {
                        "description": "Order tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "Token",
                "description": "Masukkan token Sanctum dengan format \"Bearer {token}\"."
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}